lib/arp/requirements.txt
absl-py==0.13.0
adbutils==0.9.1
altgraph==0.17.2
apkutils2==1.0.0
astunparse==1.6.3
blinker==1.4
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
cigam==0.0.3
clang==5.0
click==8.0.1
cryptography==36.0.1
decorator==4.4.2
Deprecated==1.2.11
deprecation==2.1.0
dicttoxml==1.7.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.34.1
h5py==3.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
logzero==1.6.3
lxml==4.6.2
macholib==1.15.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==20.9
Pillow==8.1.2
progress==1.5
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyelftools==0.27
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyMySQL==1.0.2
pyparsing==2.4.7
requests==2.25.1
requests-oauthlib==1.3.0
retry==0.9.2
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.23
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing-extensions==3.7.4.3
uiautomator2==2.13.0
urllib3==1.26.3
Werkzeug==2.0.1
whichcraft==0.6.1
wrapt==1.12.1
xmltodict==0.12.0
